fix(ISSUE-P1-13/15,P2-19/20/21,P3-59/61)+docs: 设备侧实测收口批次
- P1-13: 写侧 Argon2 P 改 UInt32 编码 + 写侧契约断言测试(4例)；官方 verify 工具可解本仓产物
- P1-15(新): HardenedXmlReader 兼容 Android Harmony SAX 延迟拒绝加固特性，修复明文导入 KeePass XML 设备侧全量失败 + 回归测试(3例)
- P2-19: 密码库与加密页算法/KDF/Argon2/压缩改由真实文件头下发，去除静态假值与同页矛盾；解锁页/关于页能力口径纠偏
- P2-20: 新增 SafDocumentCleanup，导出失败/取消不再残留 0 字节文件(4 路径)
- P2-21(新): 接线返回键锁定；ExtendedSettingsStore 快照上移 @Singleton 修复跨 ViewModel 域内存态分裂
- P3-59: 文件路径/默认用户名真实下发，空值「未设置」占位
- P3-61: 健康度审计行未扫描时中性「未扫描」徽标(hasScanned)
- P1-14 复核结案(行为符合设计，上会话系取证假象)；P3-60/P3-62 复核结案；P3-63 升级确认缺陷
- 新登记 P3-65(TAN/UUID 开关无持久化无消费方)、P3-66(自动填充端到端未实测)
- 全量 test --rerun-tasks: 1382 例 / 0 失败 / 0 错误 / 13 跳过；AGENTS 基线、RESOLVED_LOG §27、HANDOVER 已更新

+++ database/src/main/java/com/keepasskey/database/file/KdbxKdfParameterCodec.kt
@@ -43,7 +43,10 @@ internal object KdbxKdfParameterCodec {
             }
             is KdfParameters.Argon2 -> {
                 vd.setByteArray("S", params.salt)
-                vd.setUInt64("P", params.parallelism.toLong())
+                // ISSUE-P1-13：KDBX4 规范规定 Argon2 `P`（Parallelism）以 UInt32 写出——
+                // 官方 KeePass / KeePassXC 按严格 uint 读取，UInt64 编码会让官方客户端
+                // 以默认值派生密钥而无法解锁。I / M 保持 UInt64，V 与 P 同为 UInt32。
+                vd.setUInt32("P", params.parallelism.toLong())
                 vd.setUInt64("M", params.memoryInBytes)
                 vd.setUInt64("I", params.iterations)
                 vd.setUInt32("V", params.version.toLong())
